requirements.txt
astroid==2.2.5
certifi==2019.9.11
chardet==3.0.4
Click==7.0
colorama==0.4.1
Flask==1.1.1
idna==2.8
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.2
MarkupSafe==1.1.1
mccabe==0.6.1
pkg-resources==0.0.0
pycryptodome==3.9.0
pylint==2.3.1
PyQt5==5.13.1
PyQt5-sip==4.19.19
PyYAML==5.1.2
pyzmq==18.1.0
requests==2.22.0
six==1.12.0
toml==0.10.1
typed-ast==1.4.0
urllib3==1.25.3
Werkzeug==0.15.6
wrapt==1.11.2
